+++ examples/05-get-account-balances.ts
@@ -7,7 +7,7 @@ async function main() {
     // initializating sdk...
     const primeSdk = new PrimeSdk({ privateKey: process.env.WALLET_PRIVATE_KEY }, {
         chainId: Number(process.env.CHAIN_ID),
-        projectKey: '', // project key
+        projectKey: 'public-prime-testnet-key', // project key
     });
 
     const balances = await primeSdk.getAccountBalances({